thursday hw

+++ api/views.py
@@ -1,14 +1,17 @@
-from rest_framework.decorators import api_view
+from rest_framework.decorators import api_view, permission_classes
 from rest_framework.response import Response
+from rest_framework.permissions import IsAuthenticated
 
 #import models
 from django.apps import apps
 from rest_framework.serializers import Serializer
-from .models import Product
+from .models import Cart, Product
 Post = apps.get_model('blog', 'Post')
+from rest_framework.authtoken.models import Token
 
 #import serializers
-from .serializers import PostSerializer, ShopSerializer
+from .serializers import CartSerializer, PostSerializer, ShopSerializer, RegistrationSerializer, CreatePostSerializer
+
 
 
 # Create your views here.
@@ -31,25 +34,37 @@ def individualPost(request, post_id):
     return Response(serializer.data)
 
 @api_view(["POST"])
+@permission_classes((IsAuthenticated,))
 def createPost(request):
-    serializer = PostSerializer(data=request.data)
+    user = request.user
+    post = Post(author=user)
+    serializer = CreatePostSerializer(post,data=request.data)
     if serializer.is_valid():
         serializer.save()
-    return Response(serializer.data)
+        return Response(serializer.data)
+    return Response(serializer.errors)
 
 @api_view(["POST"])
+@permission_classes((IsAuthenticated,))
 def updatePost(request, post_id):
     post = Post.objects.get(id=post_id)
-    # if post.author.id != request.data.author:
-    #     return Response({"messages": "You cannot delete another user's post."})
+    user = request.user
+    if post.author != user:
+        return Response({'response': "You don't have permission to edit that."})
     serializer = PostSerializer(instance=post, data=request.data)
     if serializer.is_valid():
         serializer.save()
-    return Response(serializer.data)
+        return Response(serializer.data)
+    return Response(serializer.errors)
+ 
 
 @api_view(["DELETE"])
+@permission_classes((IsAuthenticated,))
 def deletePost(request, post_id):
     post = Post.objects.get(id=post_id)
+    user = request.user
+    if post.author != user:
+        return Response({'response': "You don't have permission to edit that."})
     post.delete()
     return Response({"messages": "Successfully deleted the post."})
 
@@ -65,13 +80,46 @@ def individual(request, product_id):
     serializer = ShopSerializer(product, many=False)
     return Response(serializer.data)
 
-@api_view(["GET"])
-def login(request):
-    print("Login get request went through.")
-    return Response({"messages": "Successfully logged in."})
-
-@api_view(["GET"])
+@api_view(["POST"])
 def register(request):
-    print("Register get request went through.")
-    return Response({"messages": "Successfully signed up."})
+    serializer = RegistrationSerializer(data = request.data)
+    if serializer.is_valid():
+        account = serializer.save()
+        token = Token.objects.get(user=account).key
+        data = {
+            'response': "Successfully registered a new user.",
+            'email': account.email,
+            'username': account.username,
+            'token': token,
+        }
+    else:
+        data = serializer.errors
+    return Response(data)
 
+@api_view(["POST"])
+@permission_classes((IsAuthenticated,))
+def addCart(request):
+    serializer = CartSerializer(data=request.data)
+    if serializer.is_valid():
+        serializer.save()
+        return Response("Successfully added to cart")
+    else:
+        print("addCart did not work.")
+        return Response(serializer.errors)
+
+@api_view(["GET"])
+def myCart(request,username):
+    cart = Cart.objects.filter(username = username)
+    serializer = CartSerializer(cart, many=True)
+    print("view Cart response went through")
+    return Response(serializer.data)
+
+@api_view(["DELETE"])
+@permission_classes((IsAuthenticated,))
+def deleteCart(request, username, cart_id):
+
+    cartItem = Cart.objects.get(cart_id=cart_id)
+    if cartItem.username != username:
+        return Response({"messages": "You do not have permission to delete this item."})
+    cartItem.delete()
+    return Response({"messages": "Successfully deleted the item."})
